Fix for gcc

+++ examples/ex01_mesh.cc
@@ -299,8 +299,8 @@ void TestDecimation() {
     ugu::MeshPtr src = ugu::Mesh::Create();
     ugu::Mesh dst;
     src->LoadObj(in_obj_path, data_dir);
-    ugu::QSlim(src, ugu::QSlimType::XYZ_UV, src->vertex_indices().size() * 0.1,
-               -1);
+    ugu::QSlim(src, ugu::QSlimType::XYZ_UV,
+               static_cast<int32_t>(src->vertex_indices().size() * 0.1), -1);
 
     src->WriteObj(data_dir, "plane_qslim");
   }
@@ -312,8 +312,8 @@ void TestDecimation() {
     ugu::Mesh dst;
     src->LoadObj(in_obj_path, data_dir);
 
-    ugu::QSlim(src, ugu::QSlimType::XYZ_UV, src->vertex_indices().size() * 0.1,
-               -1);
+    ugu::QSlim(src, ugu::QSlimType::XYZ_UV,
+               static_cast<int32_t>(src->vertex_indices().size() * 0.1), -1);
 
     src->WriteObj(data_dir, "spot_qslim");
   }
